Added JoyUtil method, changed DriveCommand, tweaked other things
Mostly, I added a method in JoyUtil that composed all the other drivetrain methods in there, to simplify DriveCommand so they didn't all have to be called
Then, I implemented that in DriveCommand. (The initializing/zeroing the variables in DriveCommand maybe doesn't work. The code builds but it might not actually do what it's supposed to.)
I changed/added some things in constants, and made minimal changes to RobotContainer